refactor(schemas): the query vocabularies live where both surfaces can read them (#10131)
The last 14 MCP arguments that were looser than their route, and the reason
three of them could not be fixed at all.

QUERY_ENUMS lived in src/contracts.ts, so an MCP tool could not read it --
importing contracts.ts from schemas-src/ is the dependency edge that failed
the metagraphed-data-api Workers Build twice on #10121. So the tools restated
the values, and restating an enum is how list_subnets.status came to publish a
bare {"type":"string"} while its route named `active | inactive`. It is now
schemas-src/query-enums.ts: pure literals, zero imports, readable by anything.
contracts.ts re-exports the symbol, so all 31 existing import sites are
unchanged.

Two more copies went with it:
  - schemas-src/mcp-tools/get-domain-summary.ts kept a LOCAL domain-tag list
    whose own comment said "at the time of writing" -- a list that announced it
    could drift with no way to be told when it had. Deleted; it reads
    src/domain-tags.ts, a zero-import leaf, the same way this directory has
    read src/route-limits.ts since #9127.
  - the four remaining unbounded runtime-name filters now share
    runtimeNameSchema(max), because the ceiling differs by route (100 on the
    chain-analytics feeds, 64 on the event feeds) but the meaning does not.

Also added, each replacing copies rather than adding one: reasonCodesSchema
(3 sites, 3 copies of one sentence), hashSchema (the 0x64-hex pattern), and
the bound on list_validator_economics.offset, which advertised 1,000,000
against a route that caps at 512.

THE GATE WAS COMPARING AGAINST THE WRONG ROUTE. Two tools answer for more than
one route (#9880's additionalRoutes), and validate:mcp-input-parity only ever
read the primary. list_review_gaps was reported LOOSER for naming the sort
values of the OTHER gap feed it serves -- the opposite of a defect. It now
compares against the union of every route a tool declares, and agreeing with
any of them is agreement.

TWO WRONG DESCRIPTIONS SURFACED once the enums were right, both invisible
while the parameter was a bare string:
  - list_chain_events.method said "HTTP method to use for the call" beside an
    example of `set_weights`. It is a runtime call name.
  - list_subnets.status said "rows with this health status" and gave `ok`. Its
    route accepts active | inactive. tests/mcp-input-schema.test.ts caught the
    example the moment the enum stopped being a bare string, which is that
    gate doing exactly its job.

  validate:mcp-input-parity LOOSER debt: 14 -> 0
  660 shared argument pairs: 540 identical, 89 narrowed for the context
  window, 0 looser, 31 a declared JSON-shape adaptation

openapi.json is byte-identical throughout.

Closes #10131
Refs #10064, #10060
